hola =)

+++ trunk/Cliente/src/cliente/HiloDescarga.java
@@ -23,12 +23,17 @@ public class HiloDescarga implements Runnable {
     private String nombrearchivo;
     private VentanaCliente ventana;
     private int fila;
+    private int posicion;
+    private long tamano;
+    private int accion;// variable para saber si se esta comezando a descargar por primera vez(0)
     private boolean continuar=true;
 
-    HiloDescarga(VentanaCliente ventana, String ip, String nombrearchivo) {
+    HiloDescarga(VentanaCliente ventana, String ip, String nombrearchivo,int posicionvector, int accion) {
         this.ventana = ventana;
         this.ip = ip;
         this.nombrearchivo = nombrearchivo;
+        this.posicion = posicionvector;
+        this.accion = accion;
 
     }
 
@@ -39,12 +44,9 @@ public class HiloDescarga implements Runnable {
             salida = new DataOutputStream(w.getOutputStream());
             salida.writeUTF("GET");
 
-
-
             descargararchivo();
             
             w.close();
-
 
         } catch (UnknownHostException ex) {
             System.out.println("Error de conexion");
@@ -65,17 +67,27 @@ public class HiloDescarga implements Runnable {
                 System.out.println("iniciando descargar");
                 try {
 
-                    if (f.exists()) {
-
+                    // validar que el archivo ya este descargado completo ?
+                    if ((f.exists() == true) && (ventana.getStatus(accion).matches("Cancelado") == false ))
+                    {
                         archivo = new FileOutputStream(f, true);
                         salida.writeLong(f.length());
                     } else {
                         archivo = new FileOutputStream(f);
                         salida.writeLong(0);
-
                     }
 
-                    ventana.agregarfila(nombrearchivo, Long.valueOf("0"));
+                    tamano = entrada.readLong();
+
+                    if (accion == 0)
+                      ventana.agregarfila(nombrearchivo, Long.valueOf("0"), tamano, posicion);
+                    else
+                        if(ventana.getStatus(accion).matches("Cancelado") == false )
+                          ventana.actualizarstatus("Descargando",posicion,f.length());
+                        else
+                          ventana.actualizarstatus("Descargando",posicion,Long.valueOf("0"));
+
+
                     fila = ventana.getRowCount();
 
                     byte[] buffer = new byte[1024];
@@ -84,13 +96,21 @@ public class HiloDescarga implements Runnable {
                     while ((len = entrada.read(buffer)) > 0 && continuar==true) {
                         archivo.write(buffer, 0, len);
                         total = total + len;
-                        ventana.actualizarcantidad(f.length(), fila);
-                        System.out.println("Recibiendo " + String.valueOf(len) + " total " + total);
+                        ventana.actualizarcantidad(f.length(), posicion);
+                        //System.out.println("Recibiendo " + String.valueOf(len) + " total " + total);
+                    }
+
+                    if (tamano != f.length() && continuar == true)
+                    {
+                        // aqui es que no se bajo completo y se perdio conexion con el servidor
+                        // tumbar este hilo y tratar de conectar con otro servidor.
+                        // revisar bien la parte del numero del hilo y el vector en la tabla.
+                       // (quitar el # hilo de la tabla)
+                        System.out.println("FUNDIO AQUI SE CONECTA CON OTRO");
                     }
 
 
-
-                    System.out.println("Recibido!!");
+                    System.out.println("Recibido!! QUITAR ESTE MENSAJE");
                     archivo.flush();
                     archivo.close();
                 } catch (FileNotFoundException F) {
@@ -103,10 +123,13 @@ public class HiloDescarga implements Runnable {
     }
 
     public void cerrar() throws IOException {
-
         continuar = false;
-        //w.close();
+    }
 
+    public void borrar() {
+        File f = new File(ruta + nombrearchivo);
+        f.delete();
 
     }
+
 }
